Fix various logic

+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: jin-tan <jin-tan@student.42kl.edu.my>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/16 18:04:34 by jin-tan           #+#    #+#             */
-/*   Updated: 2024/06/16 18:04:56 by jin-tan          ###   ########.fr       */
+/*   Updated: 2024/06/16 18:47:14 by jin-tan          ###   ########.fr       */
 /*                                                                            */
 /******************************************************************************/
 
@@ -18,7 +18,7 @@ void	*ft_memmove(void *dst, const void *src, size_t n)
 {
 	unsigned char *altdst;
 	unsigned char *altsrc;
-	int i;
+	size_t i;
 
 	altdst = (unsigned char *)dst;
 	altsrc = (unsigned char *)src;
